using regenerator to build, updating makefile, moving mocha options to file
