+++ test/mocha.opts
@@ -1,3 +1,4 @@
+--require co-mocha
 --reporter spec
 --timeout 10s
 --harmony-generators
